Remove color tokens from legacy theme
