+++ packages/design-tokens/utils/theme-prop-type.ts
@@ -15,10 +15,9 @@
 
 import PropTypes, { Requireable, Validator } from 'prop-types';
 
-import {
+import type {
   Theme,
   Typography,
-  Colors,
   Spacings,
   IconSizes,
   BorderRadius,
@@ -60,78 +59,6 @@ const gridPropType = PropTypes.shape({
  * ESLint rule to automatically migrate your code.
  */
 export const themePropType = PropTypes.shape({
-  colors: PropTypes.shape({
-    white: PropTypes.string.isRequired,
-    black: PropTypes.string.isRequired,
-    // Neutrals
-    n100: PropTypes.string.isRequired,
-    n200: PropTypes.string.isRequired,
-    n300: PropTypes.string.isRequired,
-    n500: PropTypes.string.isRequired,
-    n700: PropTypes.string.isRequired,
-    n800: PropTypes.string.isRequired,
-    n900: PropTypes.string.isRequired,
-    // Blues
-    b100: PropTypes.string.isRequired,
-    b200: PropTypes.string.isRequired,
-    b300: PropTypes.string.isRequired,
-    b400: PropTypes.string.isRequired,
-    b500: PropTypes.string.isRequired,
-    b700: PropTypes.string.isRequired,
-    b900: PropTypes.string.isRequired,
-    // Greens
-    g100: PropTypes.string.isRequired,
-    g200: PropTypes.string.isRequired,
-    g300: PropTypes.string.isRequired,
-    g500: PropTypes.string.isRequired,
-    g700: PropTypes.string.isRequired,
-    g900: PropTypes.string.isRequired,
-    // Violets
-    v100: PropTypes.string.isRequired,
-    v200: PropTypes.string.isRequired,
-    v300: PropTypes.string.isRequired,
-    v500: PropTypes.string.isRequired,
-    v700: PropTypes.string.isRequired,
-    v900: PropTypes.string.isRequired,
-    // Oranges
-    o100: PropTypes.string.isRequired,
-    o200: PropTypes.string.isRequired,
-    o300: PropTypes.string.isRequired,
-    o500: PropTypes.string.isRequired,
-    o700: PropTypes.string.isRequired,
-    o900: PropTypes.string.isRequired,
-    // Yellows
-    y100: PropTypes.string.isRequired,
-    y200: PropTypes.string.isRequired,
-    y300: PropTypes.string.isRequired,
-    y500: PropTypes.string.isRequired,
-    y700: PropTypes.string.isRequired,
-    y900: PropTypes.string.isRequired,
-    // Reds
-    r100: PropTypes.string.isRequired,
-    r200: PropTypes.string.isRequired,
-    r300: PropTypes.string.isRequired,
-    r500: PropTypes.string.isRequired,
-    r700: PropTypes.string.isRequired,
-    r900: PropTypes.string.isRequired,
-    // Primary
-    p100: PropTypes.string.isRequired,
-    p200: PropTypes.string.isRequired,
-    p300: PropTypes.string.isRequired,
-    p400: PropTypes.string.isRequired,
-    p500: PropTypes.string.isRequired,
-    p700: PropTypes.string.isRequired,
-    p900: PropTypes.string.isRequired,
-    // Misc
-    shadow: PropTypes.string.isRequired,
-    overlay: PropTypes.string.isRequired,
-    bodyBg: PropTypes.string.isRequired,
-    bodyColor: PropTypes.string.isRequired,
-    info: PropTypes.string.isRequired,
-    alert: PropTypes.string.isRequired,
-    confirm: PropTypes.string.isRequired,
-    notify: PropTypes.string.isRequired,
-  } as { [key in keyof Colors]: Requireable<string>['isRequired'] }).isRequired,
   spacings: PropTypes.shape({
     bit: PropTypes.string.isRequired,
     byte: PropTypes.string.isRequired,
